+++ pages/pageconnection.h
@@ -59,13 +59,13 @@ private slots:
     void on_tcpConnectButton_clicked();
     void on_udpConnectButton_clicked();
     void on_helpButton_clicked();
-    void on_canFwdButton_toggled(bool checked);
+    void on_canFwdButton_clicked();
     void on_autoConnectButton_clicked();
     void on_bleScanButton_clicked();
     void on_bleDisconnectButton_clicked();
     void on_bleConnectButton_clicked();
     void on_bleSetNameButton_clicked();
-    void on_canFwdBox_currentIndexChanged(const QString &arg1);
+    void on_canFwdBox_activated();
     void on_canRefreshButton_clicked();
     void on_canDefaultButton_clicked();
     void on_pairConnectedButton_clicked();