+++ preferences.cpp
@@ -133,6 +133,7 @@ Preferences::Preferences(QWidget *parent) :
 
     ui->uploadContentEditorButton->setChecked(mSettings.value("scripting/uploadContentEditor", true).toBool());
     ui->uploadContentFileButton->setChecked(!mSettings.value("scripting/uploadContentEditor", true).toBool());
+    ui->reconnectCanBox->setChecked(mSettings.value("reconnectLastCan", true).toBool());
 
     saveSettingsChanged();
 }
@@ -466,6 +467,9 @@ void Preferences::saveSettingsChanged()
     mLastScaling = mSettings.value("app_scale_factor", 1.0).toDouble();
     mLastIsDark = Utility::isDarkMode();
     mSettings.setValue("scripting/uploadContentEditor", ui->uploadContentEditorButton->isChecked());
+    mSettings.setValue("reconnectLastCan", ui->reconnectCanBox->isChecked());
+
+    mSettings.sync();
 }
 
 void Preferences::on_useImperialBox_toggled(bool checked)