+++ utility.h
@@ -83,8 +83,8 @@ public:
     static bool createParamParserC(ConfigParams *params, QString configName, QString filename);
     static bool createCompressedConfigC(ConfigParams *params, QString configName, QString filename);
     static uint32_t crc32c(uint8_t *data, uint32_t len);
-    static bool getFwVersionBlocking(VescInterface *vesc, FW_RX_PARAMS *params);
-    static bool getFwVersionBlockingCan(VescInterface *vesc, FW_RX_PARAMS *params, int canId);
+    static bool getFwVersionBlocking(VescInterface *vesc, FW_RX_PARAMS *params, int timeout = 4000);
+    static bool getFwVersionBlockingCan(VescInterface *vesc, FW_RX_PARAMS *params, int canId, int timeout = 4000);
     Q_INVOKABLE static bool isConnectedToHwVesc(VescInterface *vesc);
     Q_INVOKABLE static FW_RX_PARAMS getFwVersionBlocking(VescInterface *vesc);
     Q_INVOKABLE static FW_RX_PARAMS getFwVersionBlockingCan(VescInterface *vesc, int canId);